Move conversion of DNS options to addresses out of daemon struct
